[systemport]System port pvid setting error fix
Signed-off-by: vedganes <vedavinayagam.ganesan@nokia.com>

While adding system port as member for a system port channel, pvid is
set. Since pvid setting is not applicable for systemports in remote
asics, SAI returns faiure. This results in display of error.
setPortPvid() api is modified to skip setting the pvid for system ports
and an INFO is displayed.