[voq/inbandif]Code review comments fix 3
Signed-off-by: vedganes <vedavinayagam.ganesan@nokia.com>

Changes done to delete kernel neigh when there is failure in adding
kernel neigh for voq system neighbors if kernel entry already exists.
When there is mac change, system neighs are synced with changed mac.
With voq inband port type, static neighbor for this system neighbors
are added in kernel always with single chassis mac address. So to take
care of skpping adding neighbor we delete the kernel neighbor so that
next re-try will program the neighbor and hence the route addition will
proceed.